Move error list and link to wizard to its own component
