+++ src/pages/public_registration/action_accordions/SupportAccordion.tsx
@@ -1,22 +1,20 @@
-import EditIcon from '@mui/icons-material/Edit';
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
 import { LoadingButton } from '@mui/lab';
-import { Accordion, AccordionDetails, AccordionSummary, Button, Typography } from '@mui/material';
+import { Accordion, AccordionDetails, AccordionSummary, Typography } from '@mui/material';
 import { useMutation } from '@tanstack/react-query';
 import { useTranslation } from 'react-i18next';
 import { useDispatch } from 'react-redux';
-import { Link } from 'react-router-dom';
 import { createTicket, updateTicket, UpdateTicketData } from '../../../api/registrationApi';
 import { MessageForm } from '../../../components/MessageForm';
+import { RegistrationErrorActions } from '../../../components/RegistrationErrorActions';
 import { setNotification } from '../../../redux/notificationSlice';
 import { Ticket } from '../../../types/publication_types/ticket.types';
 import { Registration } from '../../../types/registration.types';
 import { isErrorStatus, isSuccessStatus } from '../../../utils/constants';
 import { dataTestId } from '../../../utils/dataTestIds';
-import { getFirstErrorTab, getTabErrors, validateRegistrationForm } from '../../../utils/formik-helpers';
-import { getRegistrationWizardPath, UrlPathTemplate } from '../../../utils/urlPaths';
+import { getTabErrors, validateRegistrationForm } from '../../../utils/formik-helpers';
+import { UrlPathTemplate } from '../../../utils/urlPaths';
 import { TicketMessageList } from '../../messages/components/MessageList';
-import { ErrorList } from '../../registration/ErrorList';
 import { TicketAssignee } from './TicketAssignee';
 
 interface SupportAccordionProps {
@@ -97,19 +95,7 @@ export const SupportAccordion = ({
             )}
 
             {isOnTasksPage && tabErrors && (
-              <div>
-                <Typography>{t('registration.public_page.error_description')}</Typography>
-                <ErrorList tabErrors={tabErrors} />
-                <Button
-                  variant="outlined"
-                  component={Link}
-                  size="small"
-                  to={`${getRegistrationWizardPath(registration.identifier)}?tab=${Math.max(getFirstErrorTab(tabErrors), 0)}`}
-                  endIcon={<EditIcon />}
-                  data-testid={dataTestId.registrationLandingPage.tasksPanel.backToWizard}>
-                  {t('registration.public_page.go_back_to_wizard')}
-                </Button>
-              </div>
+              <RegistrationErrorActions tabErrors={tabErrors} registrationIdentifier={registration.identifier} />
             )}
 
             {supportTicket.messages.length > 0 && (
