+++ src/translations/nbTranslations.json
@@ -1256,7 +1256,6 @@
       "verify_affiliation": "Identifiser tilknytning",
       "verify_contributor": "Identifiser bidragsyter"
     },
-    "curator_support_info": "Her kan du spørre din kurator om hjelp med alt som har med registrering og publisering av forskningsresultat. Du finner svar på dine spørsmål på \"Min side\" og på registreringens presentasjonsside.",
     "delete_registration": "Avpubliser forskningsresultat",
     "description": {
       "abstract": "Sammendrag",
